requirements.txt
flask
huggingface_hub
python-dotenv
# опционально для лучшего подсчёта токенов:
tiktoken